pyproject.toml
[tool.poetry]
name = "easy_bill"
version = "0.1.0"
description = ""
authors = ["Taras Stepeniuk <taras.stepeniuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
pydantic-settings = "^2.4.0"
pydantic = "^2.8.2"
uvicorn = "^0.30.5"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
alembic = "^1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"
pytest = "^8.3.2"
factory-boy = "^3.3.1"
pytest-asyncio = "^0.24.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
